requirements.txt
pandas==2.2.2
Pillow==10.4.0
streamlit==1.37.1
streamlit_drawable_canvas==0.9.3
~treamlit==1.36.0
